removed redundant casts

+++ mars/util/SystemIO.java
@@ -234,7 +234,7 @@ public class SystemIO {
         // just propagating the runtime exception (the default behavior), but 
         // I want to make it explicit.  The client needs to catch it.
         try {
-            returnValue = (int) (input.charAt(0)); // first character input
+            returnValue = input.charAt(0); // first character input
         } catch (IndexOutOfBoundsException e) // no chars present
         {
             throw e;  // was: returnValue = 0;
